+++ OUDS/Core/Components/Sources/Controls/TextInput/Internal/InputText.swift
@@ -54,10 +54,19 @@ struct InputText: View {
         .multilineTextAlignment(.leading)
         .foregroundColor(inputTextColor)
         .tint(cursorColor.color(for: colorScheme))
-        .disabled(status == .disabled || status == .readOnly || status == .loading)
+        .disabled(disabled)
     }
 
-    // MARK: - Helper
+    // MARK: - Helpers
+
+    private var disabled: Bool {
+        switch status {
+        case .disabled, .readOnly, .loading:
+            true
+        default:
+            false
+        }
+    }
 
     private var labelColor: MultipleColorSemanticToken {
         switch status {