+++ CHANGELOG.md
@@ -26,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `password input` component to version 1.3.1 (Orange-OpenSource/ouds-ios#1526)
+- `text input` component to version 1.4.1 (Orange-OpenSource/ouds-ios#1524)
 - `text area` component to version 1.2.1 (Orange-OpenSource/ouds-ios#1527)
 - themes tunings values (Orange-OpenSource/ouds-ios#1669)
 - `button` component to version 3.3.0 (Orange-OpenSource/ouds-ios#1583)